pyproject.toml
[project]
name = "event-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "cloudevents>=1.12.0",
    "dishka>=1.8.0",
    "fastapi>=0.133.1",
    "faststream[rabbit]>=0.6.6",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.10.1",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "ujson>=5.11.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.4",
]

[tool.ruff]
exclude = [
    ".env",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".venv",
]
fix = true
unsafe-fixes = true
show-fixes = true
target-version = "py311"
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
ignore = [
    "ANN401",
    "COM812",
    "D1",
    "D203",
    "D213",
    "EM",
    "FA",
    "FBT",
    "G004",
    "INP001",
    "ISC001",
    "PERF203",
    "PLR",
    "RET501",
    "RUF001",
    "S311",
    "TRY003",
    "UP040",
]
select = ["ALL"]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*.py" = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN401",
    "PLR2004",
    "S101",
    "S311",
]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder", "standard-library"]
known-third-party = []
known-local-folder = []
lines-after-imports = 2
